# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for managing expeditions, dives, and related data.",
    "version": "1.0.0",
    "title": "Expedition and Dive API"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "tags": [
    {
      "name": "Expeditions",
      "description": "API for managing expeditions"
    },
    {
      "name": "Dives",
      "description": "API for managing dives"
    },
    {
      "name": "Calendar",
      "description": "API for managing calendar entries"
    },
    {
      "name": "Camlog",
      "description": "API for managing camera log data"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/get-all-expeditions": {
      "get": {
        "tags": ["Expeditions"],
        "summary": "List all expeditions",
        "description": "Retrieve a list of all expeditions with optional sorting.",
        "parameters": [
          {
            "name": "sortfield",
            "in": "query",
            "required": false,
            "type": "string",
            "default": "ExpeditionID",
            "description": "Field to sort the expeditions by."
          },
          {
            "name": "sortorder",
            "in": "query",
            "required": false,
            "type": "string",
            "default": "ASC",
            "description": "Order to sort the expeditions (ASC or DESC)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Expedition"
              }
            }
          },
          "400": {
            "description": "Invalid sort field or order"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/get-all-dives": {
      "get": {
        "tags": ["Dives"],
        "summary": "List all dives",
        "description": "Retrieve a list of all dives with optional sorting.",
        "parameters": [
          {
            "name": "sortfield",
            "in": "query",
            "required": false,
            "type": "string",
            "default": "DiveID",
            "description": "Field to sort the dives by."
          },
          {
            "name": "sortorder",
            "in": "query",
            "required": false,
            "type": "string",
            "default": "ASC",
            "description": "Order to sort the dives (ASC or DESC)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Dive"
              }
            }
          },
          "400": {
            "description": "Invalid sort field or order"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/get-all-calendar": {
      "get": {
        "tags": ["Calendar"],
        "summary": "List all calendar entries",
        "description": "Retrieve a list of all calendar entries for expeditions and dives.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CalendarEntry"
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/get_all_camlogdata": {
      "get": {
        "tags": ["Camlog"],
        "summary": "List all camera log entries",
        "description": "Retrieve a list of all camera log entries related to dives.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CamlogEntry"
              }
            }
          },
          "404": {
            "description": "Not Found. The requested URL was not found on the server."
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/get_all_DocRickettsPilotsDiveStaging": {
      "get": {
        "tags": ["DocRicketts"],
        "summary": "Retrieve data from DocRicketts Pilots Dive Staging",
        "description": "Fetches data from the DocRicketts Pilots Dive Staging table.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DocRickettsPilotsDive"
              }
            }
          }
        }
          }

  },
  "500": {
    "description": "Server error or database connection issue",
    "schema": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    }
  },
"/adminerror": {
"get": {
"tags": ["Admin"],
"summary": "Retrieve administrative errors",
"description": "Fetches data related to administrative errors from the Bad Still Image URL table.",
"responses": {
  "200": {
    "description": "Successful operation",
    "schema": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/AdminErrors"
      }
    }
  },
  "500": {
    "description": "Server error or database connection issue",
    "schema": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    }
  }
}
}
},

"definitions": {
    "Expedition": {
      "type": "object",
      "properties": {
        "ExpeditionID": {
          "type": "integer"
        },
        "ShipName": {
          "type": "string"
        }
      }
    },
    "Dive": {
      "type": "object",
      "properties": {
        "DiveID": {
          "type": "integer"
        },
        "DiveNumber": {
          "type": "integer"
        },
        "RovName": {
          "type": "string"
        }
      }
    },
    "CalendarEntry": {
      "type": "object",
      "properties": {
        "Date": {
          "type": "string",
          "format": "date"
        },
        "Event": {
          "type": "string"
        }
      }
    },
    "CamlogEntry": {
      "type": "object",
      "properties": {
        "Timestamp": {
          "type": "string",
          "format": "date-time"
        },
        "Event": {
          "type": "string"
        },
        "Camera": {
          "type": "string"
        }
      }
    },
    "DocRickettsPilotsDive": {
      "type": "object",
      "properties": {
        "Column1": {
          "type": "string"
        },
        "Column2": {
          "type": "integer"
        }
  }
},
"DocRickettsPilotsDiveStaging": {
  "type": "object",
  "properties": {
    "Column1": {"type": "string"},
    "Column2": {"type": "integer"}
  }

  },

   "AdminErrors": {
    "type": "object",
    "properties": {
      "Column1": {"type": "string"},
      "Column2": {"type": "integer"}
    }
}
  }
}
}